test(cva): align tests with public entry points
