+++ AGENTS.md
@@ -175,9 +175,9 @@ Durable, hard-won lessons that don't fit a section above. See [Keeping this guid
 - Give deprecated value exports their own JSDoc declaration: destructuring a deprecated property into an export loses its marker in emitted declarations, even when the property and its interface are tagged. Verify editor metadata against the packed package.
 - Keep the base-only, non-composed `cx` call at fixed arity: spreading the shared empty composition array ahead of its arguments caused a repeatable throughput regression under the existing benchmark harness. Preserve the normal `cx` filtering and hook path when optimizing it.
 - Infer concatenator inputs with `readonly [...infer Inputs]`, then validate the complete callback against `string | CXInput<T>` so readonly rests and callback unions preserve the actual `cx` call contract.
-- Component-accepting guards must instantiate the class-value parameter as `any`, or narrowed components fail props contravariance; only the narrowed cases in `packages/cva/src/concatenators.test.ts` catch this, so extend them when adding a component-accepting surface. See `CXInput`'s inline comments in `src/core.ts` for its inference mechanics.
-- In `CVAComponentConfig`, gate on `[keyof Merged] extends [never]` and wrap mappings in `Uninferred`: the built-in `NoInfer` regresses literal defaults and `getSchema` assertions even with the beta package's TypeScript 6 floor. Write rejected `@ts-expect-error` configs on one line, then re-run the composition matrix in `packages/cva/src/concatenators.test.ts`; see `CVAComponentConfig`'s comment in `src/core.ts` for why.
-- `clsx/lite`'s published types lie: the `./lite` export reuses the full build's `clsx.d.mts` (claiming `ClassValue[]` inputs), but its runtime only appends string arguments and silently drops arrays/objects/numbers. Never rely on lite's types for safety; it works as a `cva/config` concatenator for string-authored components only because core flat-spreads its own arrays and passes authored values through verbatim (pinned in `packages/cva/src/concatenators.test.ts`).
+- Component-accepting guards must instantiate the class-value parameter as `any`, or narrowed components fail props contravariance; only the narrowed cases in `packages/cva/src/config.test.ts` catch this, so extend them when adding a component-accepting surface. See `CXInput`'s inline comments in `src/core.ts` for its inference mechanics.
+- In `CVAComponentConfig`, gate on `[keyof Merged] extends [never]` and wrap mappings in `Uninferred`: the built-in `NoInfer` regresses literal defaults and `getSchema` assertions even with the beta package's TypeScript 6 floor. Write rejected `@ts-expect-error` configs on one line, then re-run the composition matrix in `packages/cva/src/config.test.ts`; see `CVAComponentConfig`'s comment in `src/core.ts` for why.
+- `clsx/lite`'s published types lie: the `./lite` export reuses the full build's `clsx.d.mts` (claiming `ClassValue[]` inputs), but its runtime only appends string arguments and silently drops arrays/objects/numbers. Never rely on lite's types for safety; it works as a `cva/config` concatenator for string-authored components only because core flat-spreads its own arrays and passes authored values through verbatim (pinned in `packages/cva/src/config.test.ts`).
 - In a linked worktree (a Conductor local workspace, or any `git worktree add` checkout — `.git` is a file, not a directory), `lint-staged@17` (and likely other versions) can silently fail to commit its own fixes: the pre-commit hook's "Updating Git index again" step misresolves the `GIT_INDEX_FILE` git sets for hooks and falls back to a hardcoded `.git/index.lock`, which errors (`index file open failed: Not a directory`) because `.git` isn't a real directory there. The task (e.g. `prettier --write`) still runs and rewrites the file on disk, but the fix never gets staged, so the commit captures the pre-fix content while the working tree silently ends up with an uncommitted diff of the fix. [`.github/hooks/pre-commit`](./.github/hooks/pre-commit) works around this by unsetting `GIT_INDEX_FILE` before invoking `pnpm lint-staged`, which makes it fall back to git's own (worktree-aware) index resolution. Verified by reproducing with a real `git commit` of deliberately unformatted JSON: without the workaround, `git show HEAD:<file>` kept the raw content while the working tree had prettier's fix and `git status` showed it as modified; with the workaround, both matched and the index updated cleanly.
 - Local env-linking scripts run before Node setup, so use POSIX utilities and resolve directories with `cd -P && pwd` rather than relying on a Homebrew `realpath` binary being on PATH.
 - `pnpm test` enforces **100% coverage thresholds** (see `coverage` in [`.config/vitest.config.ts`](./.config/vitest.config.ts)) over `packages/*/src`, `test/bench/scripts`, and `.github/scripts`. New or changed code in those paths needs colocated tests in the same change or the CI `test` job fails.
