+++ CONTRIBUTING.md
@@ -25,7 +25,7 @@ focused:
 - **Commits** follow the ["Conventional Commits" specification](https://www.conventionalcommits.org/en/v1.0.0/). This allows for changelogs to be generated automatically upon release.
 - **Code** is formatted via [Prettier](https://prettier.io/)
 - **JavaScript** is written as [TypeScript](https://www.typescriptlang.org/) where possible.
-- **Tests** should live beside their source as `<source>.test.<ext>`; keep runtime and type assertions together, rather than splitting a source's unit suite by concern.
+- **Tests** should live beside their source as `<source>.test.<ext>`; in `packages/cva`, assign them to the public entry point that exposes the behavior, keep compatibility wrappers with that entry point, and keep runtime and type assertions together rather than creating thematic standalone suites.
 - **`packages/cva`'s type exports**: any type that can appear _named_ (not structurally expanded) in a consumer's generated `.d.ts` when they compile with `declaration: true` must be `export`ed from the public entry point that exposes the signature (`cva`, `cva/config`, or `cva/utils`), even if it's not meant for direct use — an unexported-but-nameable type breaks their build with a `TS4023`/`TS2459`-family error even though `cva`'s own build stays green. Not every type reachable from a public signature needs this: TypeScript structurally expands some of them (e.g. the call-signature parameter helpers) instead of naming them, so those stay unexported on purpose — export the minimum that a real consumer build fails to compile without (see `AGENTS.md`'s Learnings for how to check). These exports exist for that portability reason only, not as a feature we want people to reach for directly, so mark them with a short JSDoc saying so and don't add docs-site coverage for them. `packages/cva/src/index.test.ts` pins the current set, but only catches _losing_ one of these exports, not a new type that newly needs one.
 
 ## Getting Started
